// src/core/configuration/errors/enums/http-error-codes.ts
export enum EHttpErrorCodes {
  INTERNAL_SERVER_ERROR = 'INTERNAL_SERVER_ERROR',
  MISSING_PARAMS = 'MISSING_PARAMS',
  NOT_FOUND = 'NOT_FOUND',
  COMMUNICATION_ERROR = 'COMMUNICATION_ERROR',
  ECONNREFUSED = 'ECONNREFUSED',
  ECONNABORTED = 'ECONNABORTED',
  UNPROCESSABLE_ENTITY = 'UNPROCESSABLE_ENTITY',
  DATABASE_ERROR = 'DATABASE_ERROR',
  BAD_REQUEST = 'BAD_REQUEST',
  UNAUTHORIZED = 'UNAUTHORIZED',
  RESOURCE_ALREADY_EXISTS = 'RESOURCE_ALREADY_EXISTS',
}